requirements.txt
pytest>=7.4.0
pytest-html>=4.1.0
pytest-xdist>=3.3.0
pytest-cov>=4.1.0
allure-pytest>=2.13.0
requests>=2.31.0
pydantic>=2.8.0
python-dotenv>=1.0.0
